Intermediate commit
gen_server is not improove code. So i will not implement it.
Analysis many times copied same functions. I add module name to all
functions but it is not resolve the problem. But diagramms builds good.

It is needed change logic of collecting functions into erlout-parser. To
avoud duplicating functions they are should be added to node always.