Intermediate commit
Help file removed because it require to hardcode path to him. Text was directly added to erlang module. It is
allow to place library files in /lib and use erltoplan as utility.

Next step is addition in the makefile options "install" and "remove" that will install or remove files in the /lib.